+++ doc/src/changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## Upcoming
+## v2.3.0
+
+*Released 2025-09-01*
 
 **Features**:
 
@@ -8,7 +10,6 @@
 must evaluate to a non-empty string. If a required variable evaluates to an empty value
 then an error is reported and execution is aborted.
 ([#21](https://gitlab.com/garden-rs/garden/-/issues/21))
-
 
 - `garden ls -cc` now displays command recipes.